Fix: 쿼리 오류 수정

+++ src/main/java/UniP_server_chat/Unip_party_chat/domain/chatRoom/dto/ChatRoomsDto.java
@@ -16,6 +16,8 @@ public record ChatRoomsDto(UUID id,
                            String partyChiefName,
                            String partyChiefImageUrl,
                            Long nowCounted,
-                           Long totalCounted) {
+                           Integer totalCounted) {
+
+
 
 }
